app/requirements.txt
flask
h5py
scikit-image
numpy
pandas
torchmeta
